adjust properties type

+++ src/Data.SqlServer/Entities/Vehicle.cs
@@ -16,7 +16,7 @@ public class Vehicle
 
     public int Year { get; set; }
 
-    public float StartingBid { get; set; }
+    public double StartingBid { get; set; }
 
     public int NumberOfDoors { get; set; }
